fix: harden Method selection contracts
Resolve review findings around backend-neutral scalar signatures, detector failures, exact backend comparison, inherited authoring conflicts, and prepared backend-changing composites.

Align sklearn and TensorFlow output metadata with runtime behavior, centralize bounded Method annotation collection, expand backend override docs, and add focused regression coverage.

Verified with: ./tests.sh --ignore tests/old --ignore tests/dev -x tests

+++ tests/models/test_mnist_classifiers.py
@@ -37,7 +37,7 @@ def test_sklearn_basic_mnist_classifier_with_tfds_adapter():
     exp.train()
 
     assert categorical_accuracy(model, val_ds, batch_size=64) > 0.1
-    assert Map(val_ds, Select(0), model).spec == TensorSpec("float32", shape=(10,), backend="numpy")
+    assert Map(val_ds, Select(0), model).spec == TensorSpec("float64", shape=(10,), backend="numpy")
     assert exp.state.phase == "trained"
